feat: add authentication code and related security tests
This commit adds authentication code to check that a request has
an Authorization header in the right format, and then to validate
and decode the JWT passed as Bearer.  We also add relevant Prometheus
metrics to track failures in authorisation and token validation /
decoding.  Unit tests are included to test each stage of this and
to support this a number of mocking and helper functions have been
added.