feat: add access token parsing and scope checking authentication dependency
This commit adds an access token parsing and scope checking function that
takes a validated and decoded access token checks that it has the required
OAuth scopes, and then constructs a user object.

As part of this, and to accommodate work on authorisation, the authentication
module was renamed to register_your_data_api.authn which needed to propagate
through the security unit tests and into main.  As part of making this change
I noticed there was a missing public key testing artefact causing the unit
tests to fail, so I added this as part of the mocking rather than adding
a public key to the repo and face security warnings about leaking secrets.