test: moved key generation to generic helper
In several places we had duplicated code for generating public/private
key pairs for JWT encoding/decoding tests.  This commit moves this to
a general piece of helper code and adds a helper function to the mock
JWT Key store class to add all the keys from a dictionary to remove
duplicated code across different tests.