test: add token parser and scope checking security test
This commit adds a security test of the token parser and scope checker.
It tests 1): missing scopes are correctly detected, and 2) that the user
object is correctly setup from a parsed token.  As part of this there is
a small change to the prom metrics to remove the "missing_scope" failure
mode from JWT validation and change this to a whole new metric for
failed access control.  This is to fix behaviour where a JWT would be
both valid and invalid at the same time.