test: fix access check endpoint test bug and add failure case
Fixes a bug in the access check endpoint test where it required
a public key for the audit log in order to work.  This is changed
to programmatically generate the key in a temp file which required
a small change to allow the context class environment variable loader
to merge in os environment variables after the .env file has loaded.
This makes it easier to write test scripts that rely on programmtic
changes to env vars as they can be overwritten with a monkeypatch.
Also added a failure case for access check endpoint testing.